% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hello_I_want_to_analyse_my_PELSA_data}
\alias{hello_I_want_to_analyse_my_PELSA_data}
\title{Helping function to get started}
\usage{
hello_I_want_to_analyse_my_PELSA_data()
}
\description{
Helping function to get started
}
