% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{limma_diann_report}
\alias{limma_diann_report}
\title{Limma function}
\usage{
limma_diann_report(
  data_processed,
  conditions = c("Control", "Treatment"),
  p.adjust.method = "BH",
  p.threshold = 0.01,
  fc.threshold = log2(1.2)
)
}
\arguments{
\item{data_processed}{}

\item{conditions}{}

\item{p.adjust.method}{}

\item{p.threshold}{}

\item{fc.threshold}{}
}
\description{
Limma function
}
